windows: Don't disable sonic in local settings by default
Prior to whatever sonic_v2 is, we have to force bootSonic to false in
local settings to get the theme to load. However, it appears sonic_v2
is slowly rolling out to users and we don't need to worry about setting
this flag anymore. To fix users who are stuck in a crash-loop, remove
read-only protection from the local settings.